feat: show same column by not printing the name
