+++ go/summarize/summarize_test.go
@@ -17,13 +17,9 @@ limitations under the License.
 package summarize
 
 import (
-	"os"
+	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func tf1() readingSummary {
@@ -390,13 +386,4 @@ Line # 695
 +-------------+-----------+----------------+----------------+
 `
 	assert.Equal(t, expected, sb.String())
-}
-
-func TestSummarizeKeysFile(t *testing.T) {
-	file := readTraceFile("testdata/keys-log.json")
-	sb := &strings.Builder{}
-	printKeysSummary(sb, file, time.Date(2024, time.January, 1, 1, 2, 3, 0, time.UTC))
-	expected, err := os.ReadFile("testdata/keys-summary.md")
-	require.NoError(t, err)
-	require.Equal(t, string(expected), sb.String())
 }
