+++ go/summarize/testdata/keys-summary.md
@@ -20,26 +20,26 @@
 |Column|Position|Used %|
 |---|---|---|
 |l_orderkey|JOIN|72%|
-|l_orderkey|GROUP|17%|
+||GROUP|17%|
 |l_suppkey|JOIN|39%|
-|l_suppkey|JOIN RANGE|17%|
+||JOIN RANGE|17%|
 |l_commitdate|WHERE RANGE|28%|
 |l_receiptdate|WHERE RANGE|28%|
 |l_shipdate|WHERE RANGE|22%|
 |l_partkey|JOIN|17%|
 |l_returnflag|WHERE|6%|
 |l_shipmode|WHERE RANGE|6%|
-|l_shipmode|GROUP|6%|
+||GROUP|6%|
 
 #### Table: `orders` (11 reads and 1 writes)
 |Column|Position|Used %|
 |---|---|---|
 |o_orderkey|JOIN|83%|
-|o_orderkey|WHERE RANGE|8%|
-|o_orderkey|GROUP|8%|
+||WHERE RANGE|8%|
+||GROUP|8%|
 |o_custkey|JOIN|58%|
 |o_orderdate|WHERE RANGE|42%|
-|o_orderdate|GROUP|17%|
+||GROUP|17%|
 |o_comment|WHERE RANGE|8%|
 |o_orderpriority|GROUP|8%|
 |o_orderstatus|WHERE|8%|
@@ -51,7 +51,7 @@
 |---|---|---|
 |n_nationkey|JOIN|91%|
 |n_name|WHERE|27%|
-|n_name|GROUP|18%|
+||GROUP|18%|
 |n_regionkey|JOIN|27%|
 
 #### Table: `supplier` (8 reads and 1 writes)
@@ -66,7 +66,7 @@
 |Column|Position|Used %|
 |---|---|---|
 |c_custkey|JOIN|88%|
-|c_custkey|GROUP|38%|
+||GROUP|38%|
 |c_nationkey|JOIN|50%|
 |c_name|GROUP|25%|
 |c_acctbal|GROUP|12%|
@@ -80,21 +80,21 @@
 |---|---|---|
 |p_partkey|JOIN|67%|
 |p_brand|WHERE RANGE|17%|
-|p_brand|GROUP|17%|
+||GROUP|17%|
 |p_name|WHERE RANGE|17%|
 |p_size|WHERE RANGE|17%|
-|p_size|GROUP|17%|
+||GROUP|17%|
 |p_type|WHERE|17%|
-|p_type|WHERE RANGE|17%|
-|p_type|GROUP|17%|
+||WHERE RANGE|17%|
+||GROUP|17%|
 
 #### Table: `partsupp` (4 reads and 1 writes)
 |Column|Position|Used %|
 |---|---|---|
 |ps_suppkey|JOIN|60%|
-|ps_suppkey|WHERE RANGE|20%|
+||WHERE RANGE|20%|
 |ps_partkey|JOIN|40%|
-|ps_partkey|GROUP|40%|
+||GROUP|40%|
 
 #### Table: `region` (2 reads and 1 writes)
 |Column|Position|Used %|
